+++ sync-for-expenses/accounts.go
@@ -99,6 +99,13 @@ func (s *Accounts) Create(ctx context.Context, request operations.CreateAccountR
 			"5XX",
 		},
 	}, func() (*http.Response, error) {
+		if req.Body != nil {
+			copyBody, err := req.GetBody()
+			if err != nil {
+				return nil, err
+			}
+			req.Body = copyBody
+		}
 		return client.Do(req)
 	})
 	if err != nil {
@@ -232,6 +239,13 @@ func (s *Accounts) GetCreateModel(ctx context.Context, request operations.GetCre
 			"5XX",
 		},
 	}, func() (*http.Response, error) {
+		if req.Body != nil {
+			copyBody, err := req.GetBody()
+			if err != nil {
+				return nil, err
+			}
+			req.Body = copyBody
+		}
 		return client.Do(req)
 	})
 	if err != nil {