+++ sync-for-expenses/pushoperations.go
@@ -80,6 +80,13 @@ func (s *PushOperations) Get(ctx context.Context, request operations.GetPushOper
 			"5XX",
 		},
 	}, func() (*http.Response, error) {
+		if req.Body != nil {
+			copyBody, err := req.GetBody()
+			if err != nil {
+				return nil, err
+			}
+			req.Body = copyBody
+		}
 		return client.Do(req)
 	})
 	if err != nil {
@@ -207,6 +214,13 @@ func (s *PushOperations) List(ctx context.Context, request operations.ListPushOp
 			"5XX",
 		},
 	}, func() (*http.Response, error) {
+		if req.Body != nil {
+			copyBody, err := req.GetBody()
+			if err != nil {
+				return nil, err
+			}
+			req.Body = copyBody
+		}
 		return client.Do(req)
 	})
 	if err != nil {